# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.Task:
    properties:
      created_at:
        type: string
      due_date:
        type: string
      id:
        type: integer
      status:
        type: string
      title:
        type: string
      user_id:
        type: integer
    type: object
  routes.LoginUserF:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  routes.NewTask:
    properties:
      due_date:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  routes.RegisterUserF:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  routes.Response:
    properties:
      Message:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: olajuwonsinclair@gmail.com
    name: Sinclair
    url: https://x.com/syncc_crypt
  description: This is a note taker API .
  title: Task Manager
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login a Registered User
      parameters:
      - description: User Registration Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.LoginUserF'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/routes.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.Response'
      summary: Login a Registered User
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register User For the Task Manager
      parameters:
      - description: User Registration Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.RegisterUserF'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/routes.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.Response'
      summary: Register User For the Task Manager
      tags:
      - Auth
  /task/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing task owned by the authenticated user
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Task
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.NewTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.Response'
      security:
      - TokenAuth: []
      summary: Update a task
      tags:
      - Tasks
  /task/alltask:
    get:
      description: Returns a list of tasks belonging to the logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.Response'
      security:
      - TokenAuth: []
      summary: Get all tasks for the authenticated user
      tags:
      - Tasks
  /task/create:
    post:
      consumes:
      - application/json
      description: Allows an authenticated user to create a task
      parameters:
      - description: Task to create
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/routes.NewTask'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.Response'
      security:
      - TokenAuth: []
      summary: Create a new task
      tags:
      - Tasks
  /task/delete:
    delete:
      consumes:
      - application/json
      description: Deletes a task by its ID, only if it belongs to the authenticated
        user
      parameters:
      - description: Task ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/routes.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/routes.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.Response'
      security:
      - TokenAuth: []
      summary: Delete a task
      tags:
      - Tasks
  /task/specifictask:
    get:
      consumes:
      - application/json
      description: Retrieves all tasks for the logged-in user. You can filter by status
        using a query parameter.
      parameters:
      - description: Task status filter (e.g. 'pending', 'complete')
        in: query
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.Response'
      security:
      - TokenAuth: []
      summary: Get tasks for the authenticated user
      tags:
      - Tasks
securityDefinitions:
  TokenAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"